Refactor ConfirmEmail and ResetPasswordRequest pages to use reusable SendEmailForm component #738 #740
  - Refactor ResetPassword page to use same style and structure as the components mentioned above
  - Add toast when user successfully resets password
