+++ client/src/components/Authorization/ResetPassword.js
@@ -4,9 +4,10 @@ import { withRouter } from "react-router-dom";
 import * as accountService from "../../services/account.service";
 import { Formik, Form, Field, ErrorMessage } from "formik";
 import * as Yup from "yup";
-import clsx from "clsx";
-import { useStyles } from "./ResetPasswordRequest";
 import Button from "../Button/Button";
+import clsx from "clsx";
+import { useToast } from "../../contexts/Toast";
+import { useStyles } from "./SendEmailForm";
 
 const validationSchema = Yup.object().shape({
   password: Yup.string()
@@ -17,10 +18,11 @@ const validationSchema = Yup.object().shape({
     .oneOf([Yup.ref("password")], "Password does not match")
 });
 
-export function ResetPassword(props) {
+const ResetPassword = props => {
   const [success, setSuccess] = useState(false);
   const { token } = props.match.params;
   const classes = useStyles();
+  const toast = useToast();
 
   const handleSubmit = async ({ token, password }, { setFieldError }) => {
     const submitResponse = await accountService.resetPassword({
@@ -29,6 +31,7 @@ export function ResetPassword(props) {
     });
     if (submitResponse.data.isSuccess) {
       setSuccess(submitResponse.data.email);
+      toast.add("Your password has been reset. Please log in.");
     } else {
       setFieldError("passwordConfirm", submitResponse.data.message);
     }
@@ -39,7 +42,7 @@ export function ResetPassword(props) {
       <div className={classes.formContent}>
         {!success ? (
           <>
-            <h1>Reset Your Password</h1>
+            <h1 className={classes.pageTitle}>Reset Your Password</h1>
             <Formik
               initialValues={{
                 token,
@@ -51,47 +54,52 @@ export function ResetPassword(props) {
                 handleSubmit(values, action);
               }}
             >
-              {({ touched, errors, values }) => (
-                <Form className={classes.form}>
-                  <div className={classes.fieldGroup}>
-                    <Field
-                      type="password"
-                      value={values.password}
-                      name="password"
-                      placeholder="Password"
-                      className={clsx(
-                        classes.inputField,
-                        touched.password && errors.password ? classes.error : ""
-                      )}
-                    />
-                    <ErrorMessage
-                      name="password"
-                      component="div"
-                      className={classes.errorMessage}
-                    />
-                  </div>
-                  <div className={classes.fieldGroup}>
-                    <Field
-                      type="password"
-                      value={values.passwordConfirm}
-                      name="passwordConfirm"
-                      placeholder="Confirm Password"
-                      className={clsx(
-                        classes.inputField,
-                        touched.passwordConfirm && errors.passwordConfirm
-                          ? classes.error
-                          : ""
-                      )}
-                    />
-                    <ErrorMessage
-                      name="passwordConfirm"
-                      component="div"
-                      className={classes.errorMessage}
-                    />
-                  </div>
-                  <Button type="submit">Submit</Button>
-                </Form>
-              )}
+              {formikProps => {
+                const { touched, errors, values } = formikProps;
+                return (
+                  <Form>
+                    <div className={classes.fieldGroup}>
+                      <Field
+                        type="password"
+                        value={values.password}
+                        name="password"
+                        placeholder="Password"
+                        className={clsx(
+                          classes.inputField,
+                          touched.password && errors.password
+                            ? classes.error
+                            : ""
+                        )}
+                      />
+                      <ErrorMessage
+                        name="password"
+                        component="div"
+                        className={classes.errorMessage}
+                      />
+                      <Field
+                        type="password"
+                        value={values.passwordConfirm}
+                        name="passwordConfirm"
+                        placeholder="Confirm Password"
+                        className={clsx(
+                          classes.inputField,
+                          touched.passwordConfirm && errors.passwordConfirm
+                            ? classes.error
+                            : ""
+                        )}
+                      />
+                      <ErrorMessage
+                        name="passwordConfirm"
+                        component="div"
+                        className={classes.errorMessage}
+                      />
+                      <Button type="submit" color="colorPrimary">
+                        Submit
+                      </Button>
+                    </div>
+                  </Form>
+                );
+              }}
             </Formik>
           </>
         ) : (
@@ -108,7 +116,8 @@ export function ResetPassword(props) {
       </div>
     </div>
   );
-}
+};
+
 ResetPassword.propTypes = {
   match: PropTypes.shape({
     params: PropTypes.shape({
